_sidebar_logout now takes a cookie hash as an argument.  no longer gets the cookie using the getCookie method.
now calls OMP::FaultDB method getAssociations with the projectid lowercased.

and now for the major change:  rewrote &write_page logic. it is now much much more clear
and way easier to maintain.  But it now requires a 'show_content' query parameter or
a 'content' url parameter in order to display content. like wise for displaying output.
this will be in the POD in the next version.  With the way the feedback system now
mixes url and normal query parameters this makes much more sense than just checking
for query parameters when deciding on whether to show output or content.

svn path=/trunk/omp/perl/; revision=2882