file_fault form now has fields for specifying the fault date.  file_fault_output can now set the faultdate when submitting a fault.  does clever things like subtract a day if the time specified is in the future.
in query_fault_output the <date> tag is now only included in the xml query if days (delta) was
specified in the fault query form.  now has a drop down menu for including userids in the query
and a textfield for specifiying text strings to search for.  query_fault_form now uses the
GET method for submitting parameters instead of POST. this way the params are in the URL like with google (not to be confused with URL params) so if a user clicks the "back" button in their browser they wont have to resubmit the query.
This also means we can have HTML links that when clicked will bring up the results of a particular
query.

fault category is now a hidden field in the query_fault_form since we can't pass this as a URL param when we are using the GET method.

svn path=/trunk/omp/perl/; revision=2976