modified the arrangement of cells in the table generated by &proj_status_table.
the feedback system default page is now projecthome.pl.  Any functions that provide links to the
default page now point to this page.

removed two variable declarations in &project_home that were being redeclared anyway ($coi and $support).

set table width to %100 for the table generated by &project_home that contains the title and country flag.

&project_home now displays country flag icons for UK, INT, CA, NL, UH and JAC.

moved the location of the flag icon from the project status area to the top right corner of page
generated by &project_home

had an if condition in &project_home that was testing $comments for a true value.  the value was always true since its an array ref, so it now tests @$comments to see if it contains anything.

svn path=/trunk/omp/perl/; revision=3147