added the fault update form/page.  right now it's only useful for updating the time lost.
Now more careful when parsing the "actual time of fault."  So we can accept
1630, 16 30 and 16:30.

Added support for the new statuses.  Displays "open" status in red, other statuses appear as green, but a status such as "wont fix" probably should be displayed in red...

No longer using the close fault button.  Now we have the "change status" button
next to a popup menu of possible statuses.

No longer using a popup menu for specifying the author (User ID) when doing a query.  Now using a text field.

Now displaying the file date of a fault instead of the fault occurance date in the fault meta data portion.

svn path=/trunk/omp/perl/; revision=3595