Check if template is gzip'd, if so attempt to gunzip.
If the content appears to be gzip'd, but in fact is not, it will continue
with the original content which will most likely break when attempting to
make a science program. While this does not seem like the best solution,
it would generate exactly the same error as would have been thrown anyway.