Tweak SMU PA selection so that it uses TRACKING for HARP 2x1
I think that the SMU should always be aligned with the requested
coordinate frame regardless of what the rotator is being told.
The SMU will be able to compensate anyhow and it means that the
right thing occurs when the rotator doesn't go where we expect.

We notice this with non-rotationally symmetric patterns such as
2x1. If the rotator decides to choose a PA that is 90 degrees
from that expected, then the SMU will now be jiggling 90
degrees in error.

Since it is a friday special case 2x1 until this can be tested
properly.