acsis: Fix 2x1 rotator angles
Commit fb98fe6f4914fee7fd577aa9b9983a9a69adceb9 was incorrectr in
that the SMU will go to TRACKING regardless of what the ROTATOR
is doing. Revert that commit and fix the issue by choosing
restricting the rotator to 0 and 180 when using a 2x1 pattern.

The code explictly looks for "2x1" rather than trying to do
something clever with JCMT::SMU::Jiggle class to calculate
rotational symmetry.