scuba2: Turns out there is an upper limit to JOS_MIN
Eventually the plan is for JOS_MAX to go away but for now we do
a simple scaling to try to prevent any mishaps. The scaling is dumb
and does not attempt to sort out the optimal number of pongs.
(cherry picked from commit d0fe30121149e1f33674a8459c6ac8e50197b399)