Delay loading Astro::Coords
OMP::Config requires OMP::General which required Astro::Coords
which required Astro::SLA which required /star/lib to be in
the LD_LIBRARY_PATH.

Try to defer the Starlink dependency by moving the Astro::Coords
loading into the routine that wants Astro::Coords.

In reality we need a lightweight module that can be used by
OMP::General and OMP::Config in order to remove the circular
dependency as well as the demand for many modules to load just
to read a config file.