client: Rewrite radionet statistics application
The previous version was not correctly handling CAL data
since the CAL projects were never being retrieved. Additionally
the fault information did not handle faults filed on a night
that differed from the time of fault.

This implementation uses OMP::NightRep to do all the heavy lifting.

It could be made quite a bit faster if the initial date list was
paired down using a TimeAcct query.

Note that this version loses the ability to do more than one
project at the same time.