Use approximate apparent positions for MSB queries.
For the RADEC coordstype in the ompobs database table, the _run_query
method now constructs an Astro::Coords::Interpolated object.  This
treats the J2000 coordinates from the database as if they were
apparent coordinates.  This gives a significant improvement in the
speed of the query (approximately 20s instead of 30s for a query
of UKIDSS MSBs run on ipu) at the expense of a little accuracy.

A test with a sample of random coordinates showed that the
position error would be about half a degree by the year 2035.
(Proper motion is already not included in the database.)

However if the storeSciProg subroutine were changed to write
apparent coordinates into the database then the inaccuracy
would typically only be that incurred during a semester.

The _msb_row_to_msb_object subroutine now corrects the coordinate types
of the OMP::Info objects so that they continue to report RADEC
rather than INTERP.  This should not cause a problem as the database
does not contain coordinates of type INTERP.
(cherry picked from commit f778a8eeffda1ea163db2028c007b7970ccb9c9a)