listtargs.pl: polygons for rectangular scans.
For scan patterns boustrophedon and raster an AST polygon region
is constructed.  Had to work around the lack of bounds information
for polygons and calculate the bounds manually.

Also corrected posangle to posang so that offsets and scans at an
angle now work correctly and changed from TAN to MER projection in
the plot in order better to handle large areas.  (MOL seemed to
chop the edges off.)