Adding a database test script.
Created a directory t-db for scripts which write to the database,
and placed a basic README and Makefile inside.  Scripts placed
here should always check that they are using the devomp database.

Wrote one test store-accept.t which stores the science program
store-accept.xml into the database and performs a number of
accepts on it.  After each one the database is queried and
the script checks that the result is as expected.  The script
uses OMP::MSBDB::queryMSB with the 'remaining' constraint
enabled, so as MSBs are removed they disappear from the
query results and the script checks for this.  Note that the queryMSB
method attaches observations to the results by msbid, so checking the
number of observations associated with each MSB checks that the
msbid has been entered correctly into the ompobs table.