t-db/store-alter.t added to test alterations.
This script submits a science program and then edits the
MSBs in it, one at a time.  After each step it checks that
only the intended MSB has been changed in the database.

Also added a template for the PASSWORD module which supplies
the staff password for the test scripts.