Translator: set FTS-2 shutters for SpIterFTS2Obs.
Had to move the shutters section of dummy code into the FTS-2 section of
the fts2_config function before filling in the implementation because,
before the shutters were added to the FTS2_CONFIG XML, I had placed the
dummy shutter code at the end of the function, after the $fts2 object
had already been stored.

Note: the OT and Translator are still working in terms of FTS-2
port numbers (1 and 2).  The system might be more foolproof if they
were to use names like the 8D and 8C of the FTS2_CONFIG SHUT elements.