Translator: put FTS-2 parameters in config file.
Added length and speed to the configuration file for each
of the FTS-2 "Special Modes".  For the variable mode,
gave limits and a resolution conversion factor -- both taken
from the specifications section of the FTS-2 Wiki.  Note
that the values output by the current version of the OT
can cause these limits to be exceeded!