Translator: FTS-2 sampleTime as target total time.
After speaking to Per, it became apparent that the "integration
time" box in the OT, which generates the sampleTime parameter,
was intended to define a time during which the FTS should
scan back and forth.  (And not, as had previously been implemented,
the sample time for step and integrate mode.)

Therefore added a new parameter to the configuration file
for the step and integrate mode's sampling time, with an
initial dummy value of 0.1 seconds.  (It is unclear whether
this mode will ever be used anyway.)

Now the JOS num_cycles parameter is calculated to give
the specified integration time (rounded up to the
next integer number of scans).  I didn't divide it
over microsteps yet (as is done for stare and
dream mode).

Also adjusted the resolution factor from 3.0 to 2.5
(0.25 calculated by Per, converted from cm to mm).
(cherry picked from commit d9c2ddfb814b893bd3c302624ace4c2b33bcbbd1)