Translator: add option to divide scans into chunks.
This adds a new configuration parameter which is the maximum
duration of a "chunk" of a scan, in seconds.  It is specific
to each scan pattern, and currently only specified
for cv_daisy.  The translator uses this value to break the
observation into a number of cycles so that a heater-track
can be performed between them.