OMP::SpRegion: more efficient CmpRegion creation.
The method _build_cmpregions() has been changed to
_build_cmpregion($type) so that we only build the CmpRegions
which are really required.

Assembly of the CmpRegion is now done in a balanced binary tree
manner rather than in a linear method.  This minimizes the
depth of the CmpRegion objects, as the previous method combined
all of the regions in a chain.  AST appears to operate
much more efficiently in with this change, and so it is now
feasible to operate on large WFCAM programmes, and to use
the CmpRegion plotting method for moderate sized science programmes
where it was previously too slow.
(cherry picked from commit 4ac102da5bdfd908225f61837975b20928dfd8df)