OMP::SpRegion: option to plot in galactic coords.
The listtargs program gained a corresponding -plottingsystem
option.

Added code to convert the 'all' AST CmpRegion to galactic
coordinates in order to find the plotting bounds.  This makes
use of the new merge_regions subroutine in Starlink::ATL
since the CmpRegion must be built.

However the constant WORKAROUND_BOUNDS has been left on to
avoid AST errors finding the bounds -- this option must be
disabled in order to use the conversion routine.

Therefore added a -plottingbounds option to the listtargs
to allow the user to specify the actual bounds to use.