Translator: divide FTS-2 time by microsteps.
For other observing modes, the requested integration time
is divided by the number of positions in the microstep pattern
so that the total observation time matches that requested.
This commit updates the calculation of the number of FTS-2
scans to be performed so that it also follows this behaviour.

This should be the correct behaviour since the OT won't know
how many microsteps there are, so it could not otherwise
display a sensible estimate of the observation time.