cgi: include DQ analysis on nightly project log.
This patch is to allow a nightly data quality analysis to be
included in the project's nightly observation log, to go
just under the WVM graph.

The intention is that the analysis will be performed offline
to generate a report (HTML) and associated plots,
and these will be stored in a suitable location by UT date.

In order to restrict access to the analysis to those projects
having observations on a given night, it was necessary
to redirect the links in the report so that they load in
any images via a new wrapper script, "get_resource.pl".

Adjusted OMP::CGIPage::write_page so that if the same
subroutines are provided for "content" and "output",
and headers are diabled, it always runs in "output mode"
to prevent headers appearing.  This allows us to use
it to serve images without the headers ever appearing.