Translator: add FTS-2 specific step_time_error
For FTS-2 we currently disable some of the sub-arrays which has the
knock-on effect of slighly reducing the delays in the system which cause
the average step time to exceed the requested step time.  Therefore as
requested by Doug Johnstone (by Polycom, 2013-05-22, during FTS-2 E&C
run), allow an alternate value of step_time_error to be specified for
FTS-2 observations.  This should allow us to avoid the problem of FTS-2
scans ending too early because the steps in them end up shorter than
expected by giving a smaller error factor.