OMP::MSBDB: check for select zero or folders.
storeSciProg now specifies an option check_select_zero_or_folder
when constructing the MSB info objects to be inserted into the
database.  The MSB info function, when this option is specified,
changes the remaining count to OMP__MSB_REMOVED if the
MSB is in an or folder requesting zero items be selected.