client: make qstatus compatible with jqueue/uqueue
Importing some of Jonathan Kemp's changes to the qstatus
script.  The weather account previously had a modified version
of this script, and the aim of this commit is to eleminate
the need for that.

Added -semester and -fullday switches so that the user
can search for other semesters, and for a full 24 hour
period rather than just the normal operating hours.

Read the password from an environment variable STAFF_PASSWORD
if it exists.  This will allow cron jobs to securely provide
the password.  Then delay setting up ReadLine until it really
becomes necessary, either to ask the user for the telescope
or password.