Translator: adjust MAP_X/Y source for S-2 stares.
Change the TCS parameter from which the MAP_X and MAP_Y FITS
headers should be extracted for stare observations.  Note this
needs the accompanying change to the translator wiredir
to re-enable these headers for stare mode.

From: Craig Walther <c.walther@jach.hawaii.edu>
Date: Fri, 19 Jul 2013 15:57:09 -1000

It appears that we do not put the offset into our FITS headers when we
are doing POL-2 and FTS-2 observations. Therefore I propose that we use the
parameters MAP_X and MAP_Y to do that.

<DRAMA TASK="PTCS" PARAM="DEM_OFFSET.DC1" EVENT="START" MULT="206264.8062" />

Likewise for MAP_Y except it uses DEM_OFFSET.DC2.