Translator: no SCUBA-2 setup before pointing/focus
As discussed at the SCUBA-2/SciCom meeting of 20131015,
I checked with the TSS mailing list as to whether they
find the automatic setups before pointing and focus to
be useful.  All the reponses received so far were
that they should be removed.

This commit introduces a subroutine _not_needing_setup
which identifies scan_focus and scan_pointing observations.
This in turn causes insert_setup_obs to skip the setup
consideration machinery for these observations.