OMP::MSBDB: trap errors calculating observability
The "UTNov24.0 elements" coordinates for comet ISON in
project M13BI07 cause an error status -5 "numerical error"
from palPertel leading to fault 20131201.002.  This commit
adds an "eval" block around the source observability check
to prevent this error from causing the query to fail.
It assumes the source is not available in case of error.