OMP::MSBDB: add ability to kludge priorities
DIJ and HST requested the ability to kludge project priorities,
specifically for JLS projects being observed out of band.  The
reason is that some projects are being allowed to observe out of
their originally allocated bands, but they should be at low
priority when this happens.

For the time being I have just put a table of project patterns,
priority adjustments, and tau ranges in the Perl module itself.
The entry 'TJ01' is just a placeholder.  If this becomes more
widely used it might be worth considering moving the data to
a configuration file.