OMP::MSBDB: Set default max elevation for SCUBA-2
If a maximum elevation is not set (via a scheduling constraint)
and the MSB includes SCUBA-2 observations, then the maximum
elevation is now given a default value of 75 degrees -- as
agreed at today's JCMT meeting.

This avoids all MSBs needing to include a 75 degree limit (really
a technical restriction and not something that should necessarily
be in the science program), and should therefore make query results
more reliable as not all programs include the limit.