OMP::MSBDB: Include transit check in queries
During an MSB query, each observation in the MSB is checked
(for observability and other constraints) at the start and
end of the observation.  This commit adds an additional check
at transit, which only applies if the source transits during
the observation (rising at the start and setting at the end).