OMP::MSBDB: Add new azimuth-based ZOA constraint
This adds a new configuration parameter "zoa_azdist" which specifies
a minimum distance in azimuth from the ZOA target.  Set this initially
to 45 degrees for JCMT.  The constraint has been added as part of the
existing ZOA constraint -- so it only applies if the ZOA constraint
is enabled and the ZOA target is above elevation zero.