Add support for filling auto coordinates on fetch
An extra stage has been added to OMP::MSBDB::fetchMSB -- a call to
a new OMP::MSB::processAutoCoords method.  This checks the XML for
the presence of BASE elements containing a tleSystem element.
If any of these corresponds to the "AUTO-TLE" coordinate type
then it is replaced with a BASE element including actual coordinates.
These come from a new database module, OMP::TLEDB, which currently
just has a dummy implementation for its get_coord method to
allow for testing.