OMP::MSBDB: trap error in query's first HA calc/n
Commit d11124fbad introduced an eval block around the observability
test because errors raised by the modules performing that test
would cause the entire query to fail.  Subsequently commit
ac0faf2586 added HA calculations to the query in order to determine
whether or not a constraint test at transit is required (for the
purpose of checking whether SCUBA-2 observations exceed the
maximum elevation).  This introduced another point at which the
query can fail due to co-ordinate calculations, as happened
recently when an MSB was inserted into the database with an
observation which had TLE coordinates containing an unreasonable
epoch year parameter.  (Astro::Coords::TLE has meanwhile been
updated to reject epoch years which would result in negative
UNIX times.)

This commit wraps that new HA calculation in an eval block
to trap errors at this point.  The calculation could not simply
be moved inside the existing eval block because that is only
executed if the observability constraint is enabled.  There
are other calculations which could also fail, but if the
coordinates are bad it is quite likely that this would be
detected in the HA calculation.  It is tempting to place
the whole "$delta" loop in an eval block, but it is also the
case that if there are problems then we would like to
know about them.  Perhaps the ideal thing would be for the
MSB query to be able also to return a list of the MSBs
for which errors were encountered, preventing their consideration.