OMP::MSBDB: check TLEDB when inserting AUTO-TLEs
When storing AUTO-TLE coordinates in the ompobs table, the omptle
table is now searched for the given target.  If present then the
coordinates from the omptle table are copied to the ompobs table
(still with coordstype set to AUTO-TLE).  This commit is complicated
by the database not allowing two active statement handles, so we
must destroy the prepared ompobs insert statement handle before
querying the TLE table and then re-prepare it afterwards.