OMP::FeedbackDB: alter column order of ompfeedback
The order of the columns in the ompfeedback table changed, so
the code which inserts data into this table needs to do so in the
new order, since it inserts by position rather than by name.  The
mktable.pl script has also been updated for the new order
(and wider column type).