OMP::User: add option to return extracted email
OMP::User::extract_user_from_email now takes an additional parameter
which can be a scalar reference to which the extracted email address
should be written.  This is because extracting the email address
is a useful feature of this routine even if it then goes on to
fail to construct an OMP::User object.