OMP::DateTools: parse date in determine_semester
This method accepts either a string or Time::Piece object, and now
always converts to a Time::Piece immediately if given a string.
This prevents the private function _determine_pparc_semester also
having to handle both Time::Piece objects and strings (which is
actually its documented behavior: it specifies that it takes
a Time::Piece object).