OMP::Range: use inclusive intervals
Please see fault 20131213.003 for background information.

The behavior of Number::Interval changed in version 0.06 so that
the special case of min==max no longer causes the IncMin and IncMax
options to be ignored.  This meant that these single-value ranges could
no longer be used as they never matched anything, given that the OMP
didn't enable the IncMin or IncMax option.  We can restore the previous
behavior for OMP::Range objects by always enabling IncMin and IncMax.
This is actually more consistent with OMP::DBQuery::_querify which
translates "min" to ">=" and "max" to "<=".  This also allows us to
remove the widening of OMP::ArcQuery and OMP::ArchiveDB::Cache date
widening which was previously present for exclusive matching.