OMP::ProjAffiliationDB: add module for affiliation
For pilot science projects (JCMT semester 15A), JTD wants to be able
to track time observed by affiliation despite all the projects being
in the international queue.  Therefore adding a module to deal with
a new project affiliation table.  This is a new table rather than
an extra column in the project table to allow there to be multiple
affiliations (with associated fractions) for a given project.

Also added a script to populate the affiliation table.