contrib/findtarget: authorize via staff password
Now prompt for the staff password rather than checking file
permissions.  This is for consistency with other OMP scripts
such as client/qstatus.  It doesn't really provide any
security beyond stopping inadvertent use of the script by
non-staff, but the "findtarget" code no longer has the
database password embedded in it, so the file permission
checks are less necessary now.