CGIComponent/Obslog.pm: fix colors in 'comment on obs' html.
Previously the code attempted to use 'obslog-<state>' as a defined
html color, resulting in a very odd color choice that was not the same
as the css defined colors for obslog-bad etc.

Possibly this worked in previous browsers? Or no one noticed the
colors have always been wrong on the web page.