Translator: keep rotator PA in converted grid/frsw
Grid (including stare) frequency switched observations are converted to
1x1 jiggles but this was causing the rotator angle to be lost because
the rotator_config method then looked for jiggle mode PA parameters.

This commit introduces a new item into the "info" hash,
"isConvertedGridFreqSw", which indicates when this conversion has been
applied.  The rotator_config method now checks for this parameter when
determining whether the observation is a jiggle or a stare.

See JCMT fault: 20151205.001