Translator: allow OT to specify rotator angles
Now look for an element rotatorAngles in SpIterJiggleObs and SpIterStareObs.
If present, and not empty, write an array reference into the element summary
listing the specified values.

Then when writing the rotator configuration for ACSIS, check for the
presence of the rotatorAngles key.  If it isn't there, use 90 degrees
times the automatic symmetry choices as before.