client/obslog: Simplify loading procedure
The obslog loading time appeared to be dominated by the 1 second
sleep which allowed you to see the completed progress bar.  It
therefore seems better to do away with the progress bar and simply
load the modules in the normal manner.  Quite a lot of the OMP
modules were already just loaded via "use" statements at the top
of the script anyway.  (Some were even loaded in both places -- at
the top of the script and in the subroutine which loaded modules
and updated the progress bar.)  When there is a delay loading the
obslog, it used to happen after the progress bar had already
disappeared and it had prompted you for a user name.

"use" statements don't need to be in a BEGIN block, but I retained
BEGIN blocks around the non-"use" statements in the previous BEGIN
section of the script -- I'm not sure if it is necessary for all of
these to be in BEGIN blocks or not.

Also removed the $VERSION variable which was declared, had a
commented-out initialization and was never used.