OMP::CGIPage::User: allow apostrophes in names
When updating user account details, allow the name to include
apostrophes.  This is to accommodate names such as "O'Reilly"
which include this character.  Since there are already some
OMP accounts of users with names including this character
(the mkuser.pl script doesn't prevent their creation) it
should not be a problem to allow it here.  Otherwise it is
impossible to edit accounts for users who already have this
character in their name when trying to update one of the
other details.