client/obslog: Remove useless comment font option
The obslog had an option to specify a "comment" font.  This was
used twice in the GUI.  However in both cases the font for the
widget in question was immediately changed to the "var" font.
As a result, the "comment" font setting had no effect.  It is
therefore being removed to avoid confusion.