client/obslog: Use Tk MainWindow for obslog
Now use the actual Tk MainWindow for the primary obslog window
(as created by the create_main_window function).  This is so
that the application automatically exits when this window
is closed.  It also causes the window to be shown right away
(in a blank form) rather than only when the search results are ready.
This should be useful as the user will be able to see that the
obslog is loading, and they will be able to position the window
while waiting for the results to load.