cfg: Use special cachedir on OMP servers
Use a cachedir in "/opt/omp" rather than "/tmp" on the OMP web servers
to avoid it being placed in a private systemd temporary space.

Also use a different cachedir for the "dev" configuration, in case the
cache entries are not compatabile.  (E.g. using Time::Piece::Sybase
when we now only define Time::Piece::MySQL.)