OMP::CGIComponent::Obslog: always fetch title
The get_msb_comments function has a "get_title" option to cause it
to individually query for the MSB title, but its main call to
"historyMSBtid" would normally get the title anyway.  Now have
it return this title when "get_title" is not set.

This is so that MSB titles appear on the nightly observing report.
See fault 20180103.001.