OMP::CGIComponent::Obslog: MSB comments w/o status
This removes the observation status CSS class from MSB comments.  This
is because statuses relate to observations, not MSBs.  Also the coloring
can often be misleading -- such as an MSB starting with an MCE error in
its setup but then being completed successfully.  (Coloring the MSB
acceptance comment in this case ends up implying that it was all junk.)