OMP::MSBDB: allow checksum w/o "O" in done/undoMSB
The checksum of MSBs located within OR folders have an "O" character
appended to them.  This commit updates the doneMSB and undoMSB methods
so that, if they fail to locate the requested MSB, they also try the
same checksum with the "O" removed.  This is to suppore the case where
the second execution of an MSB has been translated before the first
execution was accepted (and hence has the old checksum, with the "O").

A test script for MSB acceptance has been added.  This tests the two
simplest cases: an MSB at the top level and a plain MSB inside an OR
folder.