OMP::MSB: share MSB-finding with OMP::SciProg
When accepting the last MSB to be chosen from an OR folder, now use
the same MSB-finding routine as OMP::SciProg.  This deals with the
"KLUGE" and "Eek" comments in OMP::MSB::hasBeenObserved and has the
benefit of improved support for survey containers, since override
targets are now taken into account.

In order to implement this, the MSB-finding part of
OMP::SciProg::locate_msbs has been extracted to make a new
subroutine _get_msbs_within_node.  (This is not a method since
an OMP::MSB object doesn't have a reference to its parent
OMP::SciProg object.)  It takes a (reference to a) hash of
arguments to be passed on to the OMP::MSB constructor.
OMP::MSB::hasBeenObserved makes use of this to skip the
fix-up operations which would normally apply on MSB object
construction.  (The intention is to form a "lightweight" MSB
object -- the only capability required being working with the
remaining counter.)