OMP::SpServer: move OT minimum version to config
Now get the minimum OT version required (and download URL) from the
configuration file rather than having them set in the module itself.
Also log the OT version used to upload programs and tidy some of the
other logging messages during program storage.

This is in preparation for the OT release for semester 18B where we
will have to update the minimum OT version in order to ensure people
do not use a mixture of versions of the software.