client: Add a new script to summarize a sci. prog.
This can be given either a project code or the name of a local XML
file containing the sciece program.  It uses a new "apply_xslt" method
of OMP::SciProg to invoke xsltproc to generate a text summary.  The
location in which stylesheets can be found is configured by a new
configuration file parameter "tomlxsltdir".