OMP::MSBDB: recalculate all checksums after accept
Accepting an MSB can cause multiple MSB checksums to changes.  Therefore
recalculate all checksums in the science program after calling
"hasBeenObserved", "undoObserve" or "hasBeenCompletelyObserved".

Also the OMP::MSB "hasBeenObserved" method no longer updates its own
checksum since this may be insufficient and we are now recalculating
all the checksums anyway.  The OMP::MSB doesn't have a reference to the
whole science program so it can't invoke the "locate_msbs" method itself.
(And if it did, it would in a way be invalidating itself, since
"locate_msbs" would make a new OMP::MSB object and update that one
instead.  That would be an odd way for an object method to behave
so it seems the only choice is to require the caller to recalculate the
checksums after invoking the method.)

Please see also fault 20180622.003.