OMP::CGIPage::Fault: simplify fault query in search
Now when searching the fault database, add new options to not retieve
the body text or list of associated projects.

The fault query could probably be further optimized by retrieving user
details in the same SQL query as the fault information.  However since
the OMP::FaultDB::_query_faultdb method already caches user information
as it is retrieved, it appears that this would make only a minor
improvement.