OMP::Info::ObsGroup: restore simple query w/o inccal
In the generalization of the "filter" method (commit 9a5fe5acf4),
OMP::Info::ObsGroup::populate started always calling "filter".
However the special calibration filtering (via the "inccal" option
of "filter") was only supposed to occur if a project identifier
was given.  Now when this is not the case, a new "anycal" option
is passed to "filter" to prevent special processing of calibrations.
In this mode, calibrations and science observations are treated
alike (as previously in the "populate" method's special case
which didn't call "filter").