OMP::CGIComponent::MSB: correct transaction inputs
The "transaction" hidden input fields in the MSB comments table
were being set up with a list of all transaction IDs for the MSB.
(So the resulting input called "transaction" would have the same
ID for all entries for the MSB and there would be several extra
hidden inputs with names and values from amongst the other IDs.)

Now this hidden input is set up for each entry in the table
where a transaction ID is present.  The "Add Comment" button is
not shown otherwise.

This is intended to address fault 20181025.001.  It should now be
possible to comment on the correct execution of an MSB.