OMP::CGIComponent:MSB: group entries by "tid"
In the MSB comments table, group the entries for each MSB by transaction
ID.  This is to ensure that additional comments added for a particular
transaction appear in the correct place.  The size of the display table
can now be determined simply from the number of groups, so the
"_count_unique_tids" function is no longer required.