OMP::MSBDoneDB: adjust handing of MSB undo feature
This updates the handling of the MSB undo feature so that the
corresponding entries in the ompmsbdone table include the transaction
ID.  This is so that an undo can be better associated with the MSB
acceptance to which it relates.

A new status OMP__DONE_UNREMOVED has been added to indicate reversal
of OMP__DONE_ALLDONE (which has been renamed to OMP__DONE_REMOVED)
in order to allow unremoval and undoing operations to be separated.
Now as well as counting each OMP__DONE_DONE as 1 MSB repeat, each
OMP__DONE_UNDONE is counted as -1 repeat so that we have a better
count of the number actually accepted on the project home page.
The observedMSBs method gained an include_undo option to support
this usage, as otherwise it would not retrieve undo events unless
also searching for additional comments/transactions.