OMP::MSBDoneDB: correct queryMSBdone comments mode
This updates the queryMSBdone comments mode so that, when it queries
for additional comments using the historyMSB method, it specifies
the project ID.  This is because checksums are not necessarily
unique between different projects.  Querying by checksum alone
returned results including some for different projects, so, depending
on the order in which the results were received, the project
ID in the returned info object could be wrong!  (And it could
contain comments for the MSB in the other project.)

The _reorganize_msb_done private method now organizes based on
project ID and checksum, since checksums aren't necessarily unique.

(cherry picked from commit 5176cab714681dddd2fd8f79d67027765adb5fa4)