OMP::MSBDoneDB: relax mult. history results check
As queryMSBdone results are no longer aggregated by checksum alone,
calls to historyMSB specifying a checksum but no project ID could
get multiple results if the same MSB appears in multiple projects.
This would raise an error (as only one result was expected).  This
commit disables the check unless a project ID is also specified.

For some cases (such as titleMSB) the multiple results don't matter
as the titles must be the same if the checksums match.  For other
uses of this method, it is best to specify a project ID to ensure
the intended MSB is selected.