filter/mail2feed: Also accept fault responses
This updates the mail2feed script to check for fault IDs as well
as project IDs in the message subject.  The function which does
this check now returns a reference to a hash of identifying information
rather than setting headers in the Mail::Audit object (which would then
have to be extracted and chomped).