OMP::CGIComponent::MSB: clear CGI obj. for MSB forms
The CGI module tries to preserve form inputs, overriding the default
values given programmatically when it already has a value for a named
parameter.  The pages which list MSB history have action buttons
such as "Undo" and "Add Comment".  Each is in a small form including
hidden fields identifying the MSB (and transaction where appropriate).
When these buttons are used, an action is performed (with the result
appearing at the top of the page) and then the MSB history is shown
again below, as part of the same request.  This means that when
the forms were generated, each contained the hidden parameters for
the MSB with which the user had just interacted.  Then any further
interactions would apply to that MSB as opposed to the (probably
different) MSB intended.

A possible solution would have been to call $q->delete_all()
to clear all of the CGI parameters before calling the
OMP::CGIComponent::MSB::msb_comments function.  However this function
itself reads the "show" CGI parameter.  (And it could cause problems
if a CGI handler wished to do something requiring the CGI parameters
after calling msb_comments.)  Therefore this function now creates a new
CGI object before generating the per-MSB (and per-transaction) forms.
The new CGI object is initialized from an empty hash to ensure that
it is empty.  Unfortunately this means all other properties of the
current CGI object are unavailable, but that may not be too much of
a problem for this function.