NightRep: don't charge for BAD or JUNK observations by default
Instead, these are put into a new $tel . 'BADOBS' dummy project. In
the nightrep.pl client code, this is not included in the final times
and is not stored to the database. It is only shown for informational
purposes, as the time of observations marked as BAD or JUNK would
normally be charged to a fault or to weather.

(cherry picked from commit 79ccd80771321a53b15716d1263373bae35d6d55)