OMP::Translator: extract _determine_best_sideband
A function to read the sideband preference file and return the best sideband
for a given sky frequency has been extracted from the "fe_config" method.
This is to separate the logic of processing the sideband information from
that used to parse the preference file, hopefully making the "fe_config"
method clearer and allowing the sideband of each subsystem to be processed
consistently even when no file is found (when support for 2SB operation is
added).