OMP::Translator: set subsystem-spec. SPW sidebands
The "fe_config" method now estimates the rest frame equivalent
of the LO frequency and uses this to determine which sideband
each subsystem is using.  If the first subsystem doesn't use the
sideband selected for the whole observation, or the mode is SSB
but subsystems use multiple sidebands, an error is thrown.

In the case of 2SB mode, individual sideband settings are added
to the subsystem frequency configuration.  (To maintain the
current behavior, these are all set to the same value for
DSB mode.)

The "spw_list" method now uses the subsystem specific sideband
values to set the "fe_sideband" attribute.